[ntuple] Fix printing of std::int8_t
Apply the same cast to int as std::uint8_t.

+++ tree/ntuple/v7/src/RFieldVisitor.cxx
@@ -226,7 +226,7 @@ void ROOT::Experimental::RPrintValueVisitor::VisitInt8Field(const RIntegralField
 {
    PrintIndent();
    PrintName(field);
-   fOutput << fValue.GetRef<std::int8_t>();
+   fOutput << static_cast<int>(fValue.GetRef<std::int8_t>());
 }
 
 void ROOT::Experimental::RPrintValueVisitor::VisitInt16Field(const RIntegralField<std::int16_t> &field)
